Added a pause feature
Created a new variable in the StateMachine to easily get a string of the name of the current state.
Created a paused variable that will be used in update so the statemachine and scrolling will only run if the game is not paused. The game can also only be paused if the current state is the 'play' state.